+++ backend/typescript/seeders/2022.10.20T21.57.52.sample-creators.ts
@@ -10,20 +10,20 @@ const seedCreators = [
     age_range: "[10, 30)",
     timezone: "EDT",
     bio: "A mysterious author who writes detective stories to unravel great conspiracies",
-    isApproved: true,
+    is_approved: true,
     createdAt: new Date(Date.now()),
     updatedAt: new Date(Date.now()),
   },
   {
     id: 2,
-    user_id: 1,
+    user_id: 2,
     location: "USA",
     rate: 2.3,
     genre: "Horror",
     age_range: "[15, 28)",
     timezone: "EDT",
     bio: "A proactive author using books to teach youngsters about the horrors in the real world like those found in abandoned ghost towns",
-    isApproved: false,
+    is_approved: false,
     createdAt: new Date(Date.now()),
     updatedAt: new Date(Date.now()),
   },
