+++ frontend/src/components/pages/AdminDashboard/ApproveConfirmationModal.tsx
@@ -10,25 +10,23 @@ import {
 } from "@chakra-ui/react";
 import React from "react";
 
-export type DeleteConfirmationProps = {
+export type ApproveConfirmationProps = {
   isOpen: boolean;
   onClose: () => void;
-  onDelete: () => Promise<void>;
-  reviewName?: string;
+  onApprove: () => Promise<void>;
   authorName?: string;
 };
 
 /**
- * DeleteConfirmationModal displays and asks for confirmation when
- * the delete button on an Admin Dashboard row is clicked
+ * ApproveConfirmationModal displays and asks for confirmation when
+ * the approve button on an Admin Dashboard row is clicked
  */
-const DeleteConfirmationModal = ({
+const ApproveConfirmationModal = ({
   isOpen,
   onClose,
-  onDelete,
-  reviewName,
+  onApprove,
   authorName,
-}: DeleteConfirmationProps): React.ReactElement => {
+}: ApproveConfirmationProps): React.ReactElement => {
   return (
     <Modal isOpen={isOpen} onClose={() => onClose()} isCentered>
       <ModalOverlay />
@@ -44,25 +42,19 @@ const DeleteConfirmationModal = ({
           Hey wait!
         </ModalHeader>
         <ModalCloseButton />
-        {reviewName ? (
-          <ModalBody ml="40px" mr="40px" fontFamily="DM Sans">
-            Are you sure you want to delete this review of {reviewName}?
-          </ModalBody>
-        ) : (
-          <ModalBody ml="40px" mr="40px" fontFamily="DM Sans">
-            Are you sure you want to delete {authorName}?
-          </ModalBody>
-        )}
+        <ModalBody ml="40px" mr="40px" fontFamily="DM Sans">
+          Are you sure you want to approve {authorName}?
+        </ModalBody>
         <ModalFooter mb="40px" mr="40px" ml="40px">
           <Button
-            w="167px"
+            w="195px"
             colorScheme="teal"
             onClick={() => {
               onClose();
-              onDelete();
+              onApprove();
             }}
           >
-            Yes, delete {reviewName ? "review" : "creator"}
+            Yes, approve creator
           </Button>
           <Button
             w="167px"
@@ -79,4 +71,4 @@ const DeleteConfirmationModal = ({
   );
 };
 
-export default DeleteConfirmationModal;
+export default ApproveConfirmationModal;
